Don't loose precision while reading folders from a SevenZFile

+++ src/main/java/org/apache/commons/compress/archivers/sevenz/Folder.java
@@ -63,7 +63,7 @@ final class Folder {
      */
     int numUnpackSubStreams;
 
-    int findBindPairForInStream(final int index) {
+    int findBindPairForInStream(final long index) {
         if (bindPairs != null) {
             for (int i = 0; i < bindPairs.length; i++) {
                 if (bindPairs[i].inIndex == index) {
